refactor: hook 내부 테스트 __test__폴더로 이동
